+++ app/src/components/ListItems/index.js
@@ -4,6 +4,7 @@ import {ListItem} from 'native-base';
 
 import styles from './styles';
 import textStyles from '../../theme/parkspotStyles';
+import {HumanizeHelper} from '../../helper/HumanizeHelper';
 
 
 export interface Props {
@@ -16,13 +17,12 @@ export interface State {
 
 }
 
-const humanizeDistance = require('../../helper/humanizeDistance');
 
 export class PlaceListItem extends Component {
 
   render() {
     const p = this.props.place;
-    // const distance = humanizeDistance(this.props.parkspot.dist);
+    // const distance = HumanizeHelper.humanizeDistance(this.props.parkspot.dist);
     return (
       <ListItem style={styles.listContainer} onPress={() => {
         this.props.onPress();